+++ app/components/shared/sidebar_component.rb
@@ -2,8 +2,9 @@
 
 module Shared
   class SidebarComponent < ViewComponent::Base
-    def initialize(title:)
-      @title = title
+    def initialize(title:, window_index:)
+      @title        = title
+      @window_index = window_index
     end
   end
 end