+++ app/models/portfolios/cash.rb
@@ -6,6 +6,12 @@ module Portfolios
 
     belongs_to :portfolio
 
+    has_many :operations,
+             class_name:  'Portfolios::Cashes::Operation',
+             inverse_of:  :cash,
+             foreign_key: :portfolios_cash_id,
+             dependent:   :destroy
+
     monetize :amount_cents
   end
 end